Clean up source directory

Find, and delete, file that were generated by running a report. Until you're comfortable with what this will do, you are strongly recommended to run orderly_cleanup_status first to see what will be deleted.

orderly_cleanup(name = NULL, root = NULL, locate = TRUE)

orderly_cleanup_status(name = NULL, root = NULL, locate = TRUE)

Arguments

name

Name of the report directory to clean (i.e., we look at src/<name> relative to your orderly root

root

The path to the root directory, or NULL (the default) to search for one from the current working directory if locate is TRUE. This function does require that the directory is configured for orderly, and not just outpack (see orderly_init for details).

locate

Logical, indicating if the configuration should be searched for. If TRUE and config is not given, then orderly looks in the working directory and up through its parents until it finds an .outpack directory

Value

An (currently unstable) object of class orderly_cleanup_status within which the element delete

indicates files that would be deleted (for orderly_cleanup_status) or that were deleted (for orderly_cleanup)

Details

After file deletion, we look through and remove all empty directories; orderly2 has similar semantics here to git where directories are never directly tracked.

For recent gert (not yet on CRAN) we will ask git if files are ignored; if ignored then they are good candidates for deletion! We encourage you to keep a per-report .gitignore that lists files that will copy into the source directory, and then we can use that same information to clean up these files after generation. Importantly, even if a file matches an ignore rule but has been committed to your repository, it will no longer match the ignore rule.

Notes for user of orderly1

In orderly1 this function has quite different semantics, because the full set of possible files is always knowable from the yaml file. So there, we start from the point of view of the list of files then compare that with the directory.