Read configuration

Read the current orderly configuration, stored within the outpack root, along with any orderly-specific extensions.

orderly_config(root = NULL, locate = TRUE)

Arguments

root

The path to the root directory, or NULL (the default) to search for one from the current working directory if locate is TRUE. This function does not require that the directory is configured for orderly, and can be any outpack root (see orderly_init for details).

locate

Logical, indicating if the root should be searched for. If TRUE, then we looks in the directory given for root (or the working directory if NULL) and then up through its parents until it finds an .outpack directory or orderly_config.yml

Value

A list of configuration options:

core: The most important options about the outpack store, containing:

path_archive: The path to the human-readable packet archive, or NULL if disabled (set in orderly_config_set as core.path_archive)

use_file_store: Indicates if a content-addressable file store is enabled (core.use_file_store)

require_complete_tree: Indicates if this outpack store requires all dependenies to be fully available (core.require_complete_tree)

hash_algorithm: The hash algorithm used (currently not modifiable)

logging: Control over logging

threshold: Level of verbosity, set via logging.threshold

console: Indicates if logging to the console is enabled by default, set via logging.console

location: Information about locations; see outpack_location_add, outpack_location_rename and outpack_location_remove to interact with this configuration, or outpack_location_list to more simply list available locations. Returns as a data.frame with columns name, id, priority, type and args, with args being a list column.

orderly: A list of orderly-specific configuration, currently this is unstable.