Copy a simple orderly example

Copy a simple orderly example for use in the docs. This function should not form part of your workflow!

Usage

orderly_example(..., names = NULL, example = "demo", dest = NULL)

Arguments

...

Arguments passed through to orderly_init()

names

Optionally, names of the reports to copy. The default is to copy all reports.

example

The name of the example to copy. Currently only "simple" and "demo" are supported.

dest

The destination. By default we use withr::local_tempdir() which will create a temporary directory that will clean itself up. This is suitable for use from the orderly examples, but you may prefer to provide your own path, in which case the path must not already exist.

Value

Invisibly, the path to the example.

Examples

path <- orderly_example()
#> ✔ Created orderly root at '/tmp/Rtmp6Hh44x/orderly2_ex_1d3e3b94b7c9'
orderly_list_src(root = path)
#> [1] "data"       "depends"    "parameters" "run_info"   "shared"    
#> [6] "strict"