Show an example file

Show a file from within one of the examples. This function exists for use within orderly help files, vignettes and tutorials and is not meant to form part of your workflows, unless you are doing something very peculiar.

Usage

orderly_example_show(name, file = NULL, example = "demo")

Arguments

name

The name of the report within the example.

file

The name of the file within the report. The default is to show the main orderly file (i.e., <name>.R)

example

The name of the example to look in. The default demo is a sprawling set of source designed to show off different orderly features.

Value

Nothing, called for its side effects only.

Details

All orderly examples here are runnable, though some will naturally have some pre-requisites (e.g., using a dependency will require that the dependency has been run first).

Examples

# We use constructions like this in the help, to show off features
# of orderly:
orderly_example_show("data")
#> 
#> ── src/data/data.R ─────────────────────────────────────────────────────────────
#> orderly_description(
#>   display = "A demo data set")
#>  
#> x <- jitter(1:30)
#> y <- 0.4 * x + 5 + rnorm(length(x), sd = 2)
#> d <- data.frame(x, y)
#>  
#> orderly_artefact("data.rds", description = "A synthetic dataset")
#> saveRDS(d, "data.rds")

# You can run this example:
path <- orderly_example()
#> ✔ Created orderly root at '/tmp/Rtmp6Hh44x/orderly2_ex_1d3e554d8c3'
orderly_run("data", root = path)
#> ℹ Starting packet 'data' `20250901-150642-7596a2e0` at 2025-09-01 15:06:42.4638
#> > orderly_description(
#> +   display = "A demo data set")
#> > x <- jitter(1:30)
#> > y <- 0.4 * x + 5 + rnorm(length(x), sd = 2)
#> > d <- data.frame(x, y)
#> > orderly_artefact("data.rds", description = "A synthetic dataset")
#> > saveRDS(d, "data.rds")
#> ✔ Finished running data.R
#> ℹ Finished 20250901-150642-7596a2e0 at 2025-09-01 15:06:42.489011 (0.02521062 secs)
#> [1] "20250901-150642-7596a2e0"