Set search options for interactive use

Set search options for interactive use of orderly; see orderly_dependency and orderly_run for details. This may be either an outpack_search_options object, or a list that will be coerced into one at the point of use (or NULL). This applies only for the current session, but applies to all interactive uses of orderly functions that might have recieved a copy of search_options via orderly_run

orderly_interactive_set_search_options(
  options = NULL,
  root = NULL,
  locate = TRUE
)

Arguments

options

Optional control over locations, when used with orderly_dependency; see of Details section of orderly_run.

root

The path to an orderly root directory, or NULL (the default) to search for one from the current working directory if locate is TRUE.

locate

Logical, indicating if the configuration should be searched for. If TRUE and config is not given, then orderly looks in the working directory and up through its parents until it finds an .outpack directory

Value

Nothing, called for its side effects