List source reports

List source reports - that is, directories within src/ that look suitable for running with orderly; these will be directories that contain an entrypoint file - a .R file with the same name as the directory (e.g., src/data/data.R corresponds to data).

Usage

orderly_list_src(root = NULL)

Arguments

root

The path to the root directory, or NULL (the default) to search for one from the current working directory. This function does require that the directory is configured for orderly, and not just outpack (see orderly_init for details).

Value

A character vector of names of source reports, suitable for passing to orderly_run

See also

Examples

path <- orderly_example()
#> ✔ Created orderly root at '/tmp/Rtmp6Hh44x/orderly2_ex_1d3e246bf716'
orderly_list_src(root = path)
#> [1] "data"       "depends"    "parameters" "run_info"   "shared"    
#> [6] "strict"