Pull one or more packets from a location

Pull one or more packets (including all their files) into this archive from one or more of your locations. This will make files available for use as dependencies (e.g., with orderly_dependency).

Usage

orderly_location_pull_packet(
  ...,
  options = NULL,
  recursive = NULL,
  root = NULL,
  locate = TRUE
)

Arguments

...

Arguments passed through to orderly_search. In the special case where the first argument is a character vector of ids and there are no named dot arguments, then we interpret this argument as a vector of ids directly. Be careful here, your query may pull a lot of data

in particular, passing NULL will match everything that every remote has!

options

Options passed to orderly_search. The option allow_remote must be TRUE as otherwise no packet could possibly be pulled, so an error is thrown if this is FALSE.

recursive

If non-NULL, a logical, indicating if we should recursively pull all packets that are referenced by the packets specified in id. This might copy a lot of data! If NULL, we default to the value given by the the configuration option require_complete_tree.

root

The path to the root directory, or NULL (the default) to search for one from the current working directory if locate is TRUE. This function does not require that the directory is configured for orderly, and can be any outpack root (see orderly_init for details).

locate

Logical, indicating if the root should be searched for. If TRUE, then we looks in the directory given for root (or the working directory if NULL) and then up through its parents until it finds an .outpack directory or orderly_config.yml

Value

Invisibly, the ids of packets that were pulled

Details

It is possible that it will take a long time to pull packets, if you are moving a lot of data or if you are operating over a slow connection. Cancelling and resuming a pull should be fairly efficient, as we keep track of files that are copied over even in the case of an interrupted pull.