Register an orderly plugin

Create an orderly plugin. A plugin is typically defined by a package and is used to extend orderly by enabling new functionality, declared in orderly_config.yml and orderly.R and affecting the running of reports primarily by creating new objects in the report environment. This system is discussed in more detail in vignette("plugins"), but will be expanded (likely in breaking ways) soon.

orderly_plugin_register(name, config, serialise, cleanup = NULL, schema = NULL)

Arguments

name

The name of the plugin, typically the package name

config

A function to read, check and process the configuration section in orderly_config.yml. This function will be passed the deserialised data from the plugin's section of orderly_config.yml, and the full path to that file. As the order of loading of plugins is not defined, each plugin must standalone and should not try and interact with other plugins at load. It should return a processed copy of the configuration data, to be passed in as the second argument to read.

serialise

A function to serialise any metadata added by the plugin's functions to the outpack metadata. It will be passed a list of all entries pushed in via orderly_plugin_add_metadata(); this is a named list with names corresponding to the field argument to orderly_plugin_add_metadata and each list element being an unnamed list with values corresponding to data.

cleanup

Optionally, a function to clean up any state that your plugin uses. You can call orderly_plugin_context from within this function and access anything you need from that. If not given, then no cleanup is done.

schema

Optionally a path to a schema for the metadata created by this plugin. See vignette("plugins") for details.

Value

Nothing, this function is called for its side effect of registering a plugin.