Prune orphan packet metadata

Prune orphan packets from your metadata store. This function can be used to remove references to packets that are no longer reachable; this could have happened because you deleted a packet manually from the archive and ran orderly_validate_archive or because you removed a location.

Usage

orderly_prune_orphans(root = NULL)

Arguments

root

The path to the root directory, or NULL (the default) to search for one from the current working directory. This function does not require that the directory is configured for orderly, and can be any outpack root (see orderly_init for details).

Value

Invisibly, a character vector of orphaned packet ids

Details

If an orphan packet is not used anywhere, then we can easily drop it - it's as if it never existed. If it is referenced by metadata that you know about from elsewhere but not locally, then that is a problem for the upstream location (and one that should not happen). If you have referenced it in a packet that you have run locally, the the metadata is not deleted.

We expose this function mostly for users who want to expunge permanently any reference to previously run packets. We hope that there should never need to really be a reason to run it.

Examples

# The same example as orderly_validate_archive; a corrupted
# archive due to the local deletion of a file
# Start with an archive containing 4 simple packets
path <- orderly_example()
#> ✔ Created orderly root at '/tmp/Rtmp6Hh44x/orderly2_ex_1d3e201ad7c1'
ids <- vapply(1:4, function(i) orderly_run("data", root = path), "")
#> ℹ Starting packet 'data' `20250901-150651-5cd83a29` at 2025-09-01 15:06:51.367126
#> > orderly_description(
#> +   display = "A demo data set")
#> > x <- jitter(1:30)
#> > y <- 0.4 * x + 5 + rnorm(length(x), sd = 2)
#> > d <- data.frame(x, y)
#> > orderly_artefact("data.rds", description = "A synthetic dataset")
#> > saveRDS(d, "data.rds")
#> ✔ Finished running data.R
#> ℹ Finished 20250901-150651-5cd83a29 at 2025-09-01 15:06:51.396139 (0.02901292 secs)
#> ℹ Starting packet 'data' `20250901-150651-6a2500c5` at 2025-09-01 15:06:51.418927
#> > orderly_description(
#> +   display = "A demo data set")
#> > x <- jitter(1:30)
#> > y <- 0.4 * x + 5 + rnorm(length(x), sd = 2)
#> > d <- data.frame(x, y)
#> > orderly_artefact("data.rds", description = "A synthetic dataset")
#> > saveRDS(d, "data.rds")
#> ✔ Finished running data.R
#> ℹ Finished 20250901-150651-6a2500c5 at 2025-09-01 15:06:51.44366 (0.02473283 secs)
#> ℹ Starting packet 'data' `20250901-150651-76df00c5` at 2025-09-01 15:06:51.468695
#> > orderly_description(
#> +   display = "A demo data set")
#> > x <- jitter(1:30)
#> > y <- 0.4 * x + 5 + rnorm(length(x), sd = 2)
#> > d <- data.frame(x, y)
#> > orderly_artefact("data.rds", description = "A synthetic dataset")
#> > saveRDS(d, "data.rds")
#> ✔ Finished running data.R
#> ℹ Finished 20250901-150651-76df00c5 at 2025-09-01 15:06:51.497709 (0.02901387 secs)
#> ℹ Starting packet 'data' `20250901-150651-84e8432f` at 2025-09-01 15:06:51.523528
#> > orderly_description(
#> +   display = "A demo data set")
#> > x <- jitter(1:30)
#> > y <- 0.4 * x + 5 + rnorm(length(x), sd = 2)
#> > d <- data.frame(x, y)
#> > orderly_artefact("data.rds", description = "A synthetic dataset")
#> > saveRDS(d, "data.rds")
#> ✔ Finished running data.R
#> ℹ Finished 20250901-150651-84e8432f at 2025-09-01 15:06:51.548392 (0.0248642 secs)
fs::file_delete(file.path(path, "archive", "data", ids[[3]], "data.rds"))

# Validate the archive ands orphan corrupt packets:
orderly_validate_archive(action = "orphan", root = path)
#> ✔ 20250901-150651-5cd83a29 (data) is valid
#> ✔ 20250901-150651-6a2500c5 (data) is valid
#> ✖ 20250901-150651-76df00c5 (data) is invalid due to its files
#> ✔ 20250901-150651-84e8432f (data) is valid

# Prune our orphans:
orderly_prune_orphans(root = path)
#> ℹ Pruning 1 orphan packet