Format outpack query for displaying to users. It will typically be easier to use the format method of orderly_query objects.

Format outpack query for displaying to users. It will typically be easier to use the format method of orderly_query objects.

orderly_query_format(query, subquery = NULL, envir = NULL)

Arguments

query

The outpack query to print

subquery

Optionally a named list of subqueries - if given, each must be a valid deparseable subquery (i.e., a literal or language object). Note that you must not provide an already-deparsed query here or it will get quoted!

envir

Optionally an environment of values to substitute in for environment: lookups; if given, then formatting a query will turn something like latest(parameter:x == environment:a) into latest(parameter:x == 1) (if a within envir is 1). This can be used to make queries self contained even after the environment has gone.

Value

Query expression as a string

Examples

orderly2::orderly_query_format(quote(name == "example"))
#> [1] "name == \"example\""
orderly2::orderly_query_format(
  quote(usedby({A})),
  subquery = list(A = quote(latest(name == "a"))))
#> [1] "usedby({latest(name == \"a\")})"

orderly2::orderly_query_format(
  quote(parameter:x == environment:x))
#> [1] "parameter:x == environment:x"
orderly2::orderly_query_format(
  quote(parameter:x == environment:x), envir = list2env(list(x = 1)))
#> [1] "parameter:x == 1"

format(orderly2::orderly_query("latest", name = "a"))
#> [1] "latest(name == \"a\")"