Run a report

Run a report. This will create a new directory in drafts/<reportname>, copy your declared resources there, run your script and check that all expected artefacts were created.

orderly_run(
  name,
  parameters = NULL,
  envir = NULL,
  logging_console = NULL,
  logging_threshold = NULL,
  search_options = NULL,
  root = NULL,
  locate = TRUE
)

Arguments

name

Name of the report to run

parameters

Parameters passed to the report. A named list of parameters declared in the orderly.yml. Each parameter must be a scalar character, numeric, integer or logical.

envir

The environment that will be used to evaluate the report script; by default we use the global environment, which may not always be what is wanted.

logging_console

Optional logical, passed through to outpack_packet_start to control printing logs to the console, overriding any default configuration set at the root.

logging_threshold

Optional logging threshold, passed through to outpack_packet_start to control the amount of detail in logs printed during running, overriding any default configuration set at the root. If given, must be one of info, debug or trace (in increasing order of verbosity). Practically this has no effect at present as we've not added any fine-grained logging.

search_options

Optional control over locations, when used with orderly_dependency; see Details.

root

The path to an orderly root directory, or NULL (the default) to search for one from the current working directory if locate is TRUE.

locate

Logical, indicating if the configuration should be searched for. If TRUE and config is not given, then orderly looks in the working directory and up through its parents until it finds an .outpack directory

Value

The id of the created report (a string)

Locations used in dependency resolution

If your packet depends on other packets, you will want to control the locations that are used to find appropriate packets. The control for this is passed through this function and not as an argument to orderly_dependency because this is a property of the way that a packet is created and not of a packet itself; importantly different users may have different names for their locations so it makes little sense to encode the location name into the source code. Alternatively, you want to use different locations in different contexts (initial development where you want to include local copies packets as possible dependencies vs resolving dependencies only as they would be resolved on one of your locations!

Similarly, you might want to include packets that are known by other locations but are not currently downloaded onto this machine - pulling these packets in could take anything from seconds to hours depending on their size and the speed of your network connection (but not pulling in the packets could mean that your packet fails to run).

To allow for control over this you can pass in an argument search_options, which is a outpack_search_options object, and allows control over the names of the locations to use, whether metadata should be refreshed before we pull anything and if packets that are not currently downloaded should be considered candidates.

This has no effect when running interactively, in which case you can specify the search options (root specific) with orderly_interactive_set_search_options

Equivalence to the old use_draft option

The above location handling generalises orderly (v1)'s old use_draft option, in terms of the location argument to orderly2::outpack_search_options`:

use_draft = TRUE is location = "local"

use_draft = FALSE is location = c(...) where you should provide all locations except local (setdiff(orderly2::outpack_location_list(), "local"))

use_draft = "newer" is location = NULL

(this last option was the one most people preferred so is the new default behaviour). In addition, you could resolve dependencies as they currently exist on production right now with the options:

location = "production", pull_metadata = TRUE, require_unpacked = FALSE

which updates your current metadata from production, then runs queries against only packets known on that remote, then depends on them even if you don't (yet) have them locally. This functionality was never available in orderly version 1, though we had intended to support it.