Information about currently running report

Fetch information about the actively running report. This allows you to reflect information about your report back as part of the report, for example embedding the current report id, or information about computed dependencies. This information is in a slightly different format to orderly version 1.x and does not (currently) include information about dependencies when run outside of orderly_run, but this was never reliable previously.

Usage

orderly_run_info()

Value

A list with elements

name: The name of the current report

id: The id of the current report, NA if running interactively

root: The orderly root path

depends: A data frame with information about the dependencies (not available interactively)

index: an integer sequence along calls to orderly_dependency

name: the name of the dependency

query: the query used to find the dependency

id: the computed id of the included packet

filename: the file used from the packet

as: the filename used locally