Copy shared resources into a packet directory

Copy shared resources into a packet directory. You can use this to share common resources (data or code) between multiple packets. Additional metadata will be added to keep track of where the files came from. Using this function requires the shared resources directory shared/ exists at the orderly root; an error will be raised if this is not configured when we attempt to fetch files.

Usage

orderly_shared_resource(...)

Arguments

...

The shared resources to copy. If arguments are named, the name will be the destination file while the value is the filename within the shared resource directory.

You can use a limited form of string interpolation in the names of this argument; using ${variable} will pick up values from envir and substitute them into your string. This is similar to the interpolation you might be familiar with from glue::glue or similar, but much simpler with no concatenation or other fancy features supported.

Value

Invisibly, a data.frame with columns here (the filenames as as copied into the running packet) and there (the filenames within shared/). Do not rely on the ordering where directory expansion was performed.

Examples

# An example in context within the orderly examples:
orderly_example_show("shared")
#> 
#> ── src/shared/shared.R ─────────────────────────────────────────────────────────
#> # Pull in the file 'shared/palette.R' as 'cols.R'
#> orderly_shared_resource(cols.R = "palette.R")
#>  
#> # Then source it, as usual
#> source("cols.R")
#>  
#> # And use the function 'palette()' found within
#> png("volcano.png")
#> image(volcano, col = palette())
#> dev.off()

# Here's the directory structure for this example:
path <- orderly_example(names = "shared")
#> ✔ Created orderly root at '/tmp/Rtmp6Hh44x/orderly2_ex_1d3e15482783'
fs::dir_tree(path)
#> /tmp/Rtmp6Hh44x/orderly2_ex_1d3e15482783
#> ├── orderly_config.yml
#> ├── shared
#> │   └── palette.R
#> └── src
#>     └── shared
#>         └── shared.R

# We can run this packet:
orderly_run("shared", root = path)
#> ℹ Starting packet 'shared' `20250901-150654-8b6449fa` at 2025-09-01 15:06:54.548933
#> > # Pull in the file 'shared/palette.R' as 'cols.R'
#> > orderly_shared_resource(cols.R = "palette.R")
#> > # Then source it, as usual
#> > source("cols.R")
#> > # And use the function 'palette()' found within
#> > png("volcano.png")
#> > image(volcano, col = palette())
#> > dev.off()
#> agg_record_896085249 
#>                    2 
#> ✔ Finished running shared.R
#> ℹ Finished 20250901-150654-8b6449fa at 2025-09-01 15:06:54.593581 (0.04464841 secs)
#> [1] "20250901-150654-8b6449fa"

# In the final archive version of the packet, 'cols.R' is copied
# over from `shared/`, so we have a copy of the version of the code
# that was used in the analysis
fs::dir_tree(path)
#> /tmp/Rtmp6Hh44x/orderly2_ex_1d3e15482783
#> ├── archive
#> │   └── shared
#> │       └── 20250901-150654-8b6449fa
#> │           ├── cols.R
#> │           ├── shared.R
#> │           └── volcano.png
#> ├── draft
#> │   └── shared
#> ├── orderly_config.yml
#> ├── shared
#> │   └── palette.R
#> └── src
#>     └── shared
#>         └── shared.R