Copy files from a packet

Copy files from a packet to anywhere. Similar to outpack_packet_use_dependency except that this is not used in an active packet context. You can use this function to pull files from an outpack root to a directory outside of the control of outpack, for example.

outpack_copy_files(
  id,
  files,
  dest,
  allow_remote = FALSE,
  overwrite = TRUE,
  envir = parent.frame(),
  root = NULL
)

Arguments

id

Optionally, an outpack id via outpack_id. If not given a new id will be generated.

files

Files to copy from the other packet. This can be (1) a character vector, in which case files are copied over without changing their names, (2) a named character vector, in which case the name will be used as the destination name, or (3) a data.frame (including tbl_df, or data.frame objects) containing columns from and to, in which case the files from will be copied with names to.

In all cases, if you want to import a directory of files from a packet, you must refer to the source with a trailing slash (e.g., c(here = "there/")), which will create the local directory here/... with files from the upstream packet directory there/. If you omit the slash then an error will be thrown suggesting that you add a slash if this is what you intended.

You can use a limited form of string interpolation in the names of this argument; using ${variable} will pick up values from envir and substitute them into your string. This is similar to the interpolation you might be familiar with from glue::glue or similar, but much simpler with no concatenation or other fancy features supported.

Note that there is an unfortunate, but (to us) avoidable inconsistency here; interpolation of values from your environment in the query is done by using environment:x and in the destination filename by doing ${x}.

dest

The directory to copy into

allow_remote

Logical, indicating if we should attempt to retrieve the file from any remote location if it cannot be found locally. If the file is large, this may take some time depending on the speed of the connection. If you use a file store, note that this does add the downloaded file into your file store, though associated with no packet so that it is subject to garbage collection (once we write support for that).

overwrite

Overwrite files at the destination; this is typically what you want, but set to FALSE if you would prefer that an error be thrown if the destination file already exists.

envir

An environment into which string interpolation may happen (see outpack_packet_use_dependency for details on the string interpolation). The default here is to use the calling environment, which is typically reasonable, but may need changing in programmatic use.

root

The outpack root. Will be searched for from the current directory if not given.

Value

Nothing, invisibly. Primarily called for its side effect of copying files from a packet into the directory dest

Details

There are different ways that this might fail (or recover from failure):

if id is not known in the metadata store (not known because it's not unpacked but also not known to be present in some other remote) then this will fail because it's impossible to resolve the files. Consider refreshing the metadata with outpack_location_pull_metadata to refresh this.

if the id is not unpacked and no local copy of the files referred to can be found, we error by default (but see the next option). However, sometimes the file you refer to might also be present because you have downloaded a packet that depended on it, or because the content of the file is unchanged because from some other packet version you have locally.

if the id is not unpacked, there is no local copy of the file and if allow_remote is TRUE we will try and request the file from whatever remote would be selected by outpack_location_pull_packet for this packet.

Note that empty directories might be created on failure.