Create outpack root

Create a new outpack root.

outpack_init(
  root,
  path_archive = "archive",
  use_file_store = FALSE,
  require_complete_tree = FALSE,
  logging_console = TRUE,
  logging_threshold = "info"
)

Arguments

root

Path to use. This path may exist, but it is an error to call this on a path that has already been initialised.

path_archive

Path to the archive directory, used to store human-readable copies of packets. If NULL, no such copy is made, and file_store must be TRUE

use_file_store

Logical, indicating if we should use a content-addressable file-store as the source of truth for packets. If archive is non-NULL, the file-store will be used as the source of truth and the duplicated files in archive exist only for convenience.

require_complete_tree

Logical, indicating if we require a complete tree of packets. This currently affects outpack_location_pull_packet, by requiring that it always operates in recursive mode. This is FALSE by default, but set to TRUE if you want your archive to behave well as a location; if TRUE you will always have all the packets that you hold metadata about.

logging_console

Logical, indicating if we should log to the console. If TRUE, then many operations will produce informational output; set to FALSE to prevent this.

logging_threshold

The degree of verbosity; one of info, debug or trace in increasing order of verbosity.

Value

Invisibly, an outpack_root object; these will change in future versions!