Pull a single packet from a location

Pull a packet (all files) from a location into this archive. This will make files available for use as dependencies (e.g., with orderly_dependency)

outpack_location_pull_packet(
  id,
  location = NULL,
  recursive = NULL,
  root = NULL,
  locate = TRUE
)

Arguments

id

The id of the packet(s) to pull

location

Control the location that the packet can be pulled from. The default (NULL) will try and pull the packet from anywhere it can be found, starting with locations that have the highest priority. Provide a string to limit the search to a particular location, or provide a number to limit to locations with at least this priority.

recursive

If non-NULL, a logical, indicating if we should recursively pull all packets that are referenced by the packets specified in id. This might copy a lot of data! If NULL, we default to the value given by the the configuration option require_complete_tree.

root

The path to the root directory, or NULL (the default) to search for one from the current working directory if locate is TRUE. This function does not require that the directory is configured for orderly, and can be any outpack root (see orderly_init for details).

locate

Logical, indicating if the root should be searched for. If TRUE, then we looks in the directory given for root (or the working directory if NULL) and then up through its parents until it finds an .outpack directory or orderly_config.yml

Value

Invisibly, the ids of packets that were pulled

Details

The behaviour of this function will vary depending on whether or not the destination outpack repository (i.e., root) uses a file store or not. If it does, then we simply import the unknown files into the store, and this will always be fairly efficient. If no file store is used then for the time being we pull all files from the upstream location, even if this means copying a file we already know about elsewhere in the outpack archive. We will improve this in a future version.