Mark files during packet run

Mark file within an in-progress packet. This will store the hash of the file within the internal outpack structures and force an error if the file is changed or deleted later. The function outpack_packet_file_list() will report on which files are marked (or unmarked) within the directory.

outpack_packet_file_mark(packet, files, status)

outpack_packet_file_list(packet)

Arguments

packet

Optionally, an explicitly-passed packet; see outpack_packet_start() for details.

files

A character vector of relative paths

status

A status to mark the file with. Must be "immutable" or "ignored"

Value

Depending on function

outpack_packet_file_mark returns nothing

outpack_packet_file_list returns a data.frame with columns path and status (immutable, ignored or unknown)