Query outpack's database

Evaluate a query against the outpack database, returning a vector of matching packet ids. Note that by default this only searches through packets that are unpacked and available for direct use on this computer; to search within packets known to other locations (and that we might know about via their metadata) you will need to use the options argument.

outpack_search(
  ...,
  parameters = NULL,
  envir = parent.frame(),
  options = NULL,
  root = NULL
)

Arguments

...

Arguments passed through to outpack_query, perhaps just a query expression

parameters

Optionally, a named list of parameters to substitute into the query (using the this: prefix)

envir

Optionally, an environment to substitute into the query (using the environment: prefix). The default here is to use the calling environment, but you can explicitly pass this in if you want to control where this lookup happens.

options

Optionally, a outpack_search_options object for controlling how the search is performed, and which packets should be considered in scope. If not provided, default options are used (i.e., orderly2::outpack_search_options())

root

The outpack root. Will be searched for from the current directory if not given.

Value

A character vector of matching ids. In the case of no match from a query returning a single value (e.g., latest(...)

or single(...)) this will be a character missing value (NA_character_)