Packet search options

Options for controlling how packet searches are carried out, for example via outpack_search and outpack_packet_use_dependency. The details here are never included in the metadata alongside the query (that is, they're not part of the query even though they affect it).

outpack_search_options(
  location = NULL,
  allow_remote = FALSE,
  pull_metadata = FALSE
)

Arguments

location

Optional vector of locations to pull from. We might in future expand this to allow wildcards, exceptions, or numeric values corresponding to the location priority (and then it's possible we'll change the name).

allow_remote

Logical, indicating if we should allow packets to be found that are not currently unpacked (i.e., are known only to a location that we have metadata from). If this is TRUE, then in conjunction with outpack_packet_use_dependency you might pull a large quantity of data.

pull_metadata

Logical, indicating if we should pull metadata immediately before the search. If location is given, then we will pass this through to outpack_location_pull_metadata to filter locations to update. If pulling many packets in sequence, you will want to update this option to FALSE after the first pull.

Value

An object of class outpack_search_options which should not be modified after creation (but see note about pull_metadata)