Run a report

Run a report. This will create a new directory in drafts/<reportname>, copy your declared resources there, run your script and check that all expected artefacts were created.

orderly_run(name, parameters = NULL, envir = NULL, root = NULL, locate = TRUE)

Arguments

name

Name of the report to run

parameters

Parameters passed to the report. A named list of parameters declared in the orderly.yml. Each parameter must be a scalar character, numeric, integer or logical.

envir

The environment that will be used to evaluate the report script; by default we use the global environment, which may not always be what is wanted.

root

The path to an orderly root directory, or NULL (the default) to search for one from the current working directory if locate is TRUE.

locate

Logical, indicating if the configuration should be searched for. If TRUE and config is not given, then orderly looks in the working directory and up through its parents until it finds an .outpack directory

Value

The id of the newly created report

Differences with orderly

There is no longer a "commit" phase, we will write a vignette explaining this change more fully.

Not supported yet:

automatic handling of README files

dependencies (at least mrc-3113)

global resources (requires more orderly2 configuration)

fields (requires more orderly2 configuration)

tags (we will probably remove these)

environment

secrets

all database support (connection, data and views, plus interpolation of parameters into queries)

automatic installation of missing packages (VIMC-6626)

check balance of the device and conneciton stack (VIMC-6628)

store random number state in metadata (VIMC-6627)

store git status in metadata (mrc-3380)

save metadata on failure (mrc-3379)

loading environment variables from orderly_envir.yml

echoing run output to console not supported (mrc-3381)

Smaller behavioural changes that might be up for changing later

Strip leading src/ from name

Allow name = NULL and use current working directory

The .outpack/ directory now marks the root, not the orderly_config.yml file, though this must also exist.