Query index

Class for managing the active index whilst evaluating a query

Public fields

index

The packet index

depends

Named list of data frames. Names are packet ids, values are packets depended on by this packet id (i.e. its parents).

uses

Named list of data frames. Names are packet ids, values are packets which are used by this packet id (i.e. its children).

root

The outpack root object

Methods

Public methods

query_index$new()

query_index$get_packet_depends()

query_index$get_packet_uses()

Method new()

Create a new query_index object

Usage

query_index$new(root, index, depends, uses)

Arguments

root

The outpack root object

index

The packet index as a data frame

depends

Named list of data frames. Names are packet ids, values are packets depended on by this packet id (i.e. its parents).

uses

Named list of data frames. Names are packet ids, values are packets used by on by this packet id (i.e. its children). This is the same data as depends but relationships flow in the other direction.

Method get_packet_depends()

Get the ids of packets which this packet depends to a specified level

Usage

query_index$get_packet_depends(id, depth)

Arguments

id

The id of the packet to get parents of

depth

Depth of parents to get, depth 1 gets immediate parents depth 2 gets parents and parents of parents, depth Inf will recurse the whole tree to get all parents

Returns

The ids of the parents of this packet

Method get_packet_uses()

Get the ids of packets which are used by this packet

Usage

query_index$get_packet_uses(id, depth)

Arguments

id

The id of the packet to get children of

depth

Depth of children to get, depth 1 gets immediate children depth 2 gets children and children of children, depth Inf will recurse the whole tree to get all children

Returns

The ids of the children of this packet